python/requirements.txt
# Framework API
fastapi>=0.95.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.25.0
pydantic>=2.0.0,<2.5.0

# Obsługa Docker
docker>=6.0.0,<6.2.0
docker-compose>=1.29.0,<1.30.0

# Bezpieczeństwo
python-multipart>=0.0.5,<0.1.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
pyjwt>=2.8.0,<2.9.0

# Bazy danych i ORM
sqlalchemy>=2.0.0,<2.1.0
asyncpg>=0.27.0,<0.30.0
aiosqlite>=0.19.0,<0.20.0

# Obsługa języków programowania
pylint>=2.17.0,<3.1.0
black>=23.9.0,<23.11.0
isort>=5.12.0,<5.13.0
mypy>=1.6.0,<1.8.0

# Wizualizacja i przetwarzanie danych
matplotlib>=3.7.0,<3.8.0
numpy>=1.24.0,<1.25.0
pandas>=2.0.0,<2.2.0
pillow>=10.0.0,<10.2.0
plotly>=5.17.0,<5.19.0
seaborn>=0.12.0,<0.14.0

# Narzędzia pomocnicze
python-dotenv>=1.0.0,<1.1.0
pyyaml>=6.0.0,<6.1.0
requests>=2.31.0,<2.32.0
aiohttp>=3.8.0,<3.9.0

# Monitoring i logowanie
prometheus-client>=0.17.0,<0.18.0
python-json-logger>=2.0.7,<2.1.0

# Cache
redis>=5.0.0,<5.1.0
cachetools>=5.3.0,<5.4.0

# Websockets
websockets>=11.0.0,<12.1.0
python-socketio>=5.9.0,<5.11.0

# CLI
click>=8.1.0,<8.2.0
typer>=0.9.0,<0.10.0

# Optymalizacja
numba>=0.57.0,<0.59.0
